Added is_on_curve func

+++ src/e521.rs
@@ -133,8 +133,8 @@ impl Point {
         let x = &self.x;
         let y = &self.y;
 
-        let x_squared = x.pow(2u64) % &p;
-        let y_squared = y.pow(2u64) % &p;
+        let x_squared = x.pow(2 as u64) % &p;
+        let y_squared = y.pow(2 as u64) % &p;
 
         let left_side = (x_squared.clone() + y_squared.clone()) % &p;
         let right_side = (BigInt::one() + d * x_squared * y_squared) % &p;
